@@ -1,6 +1,6 @@
 package com.lin.jiang.app.view;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.lin.jiang.app.R;
